refactor(wiki): break the sources panel/rail import cycle, neutral class name
Shared source helpers move to a sources.ts leaf module (the
commentCards.tsx precedent), so SourcesPanel and SourceAnchorRail no
longer import each other, and the map gains the hubspot and braintrust
icons the live ingest traffic uses. The comments-anchored class renames
to panel-anchored since it now governs both anchored panels, and the
rail dedupes its row type and open-state predicate.
