+++ frontend/src/components/wiki/SourcesPanel.tsx
@@ -9,34 +9,18 @@ import {
   Text,
 } from "@onyx-ai/opal/components";
 import { Section } from "@onyx-ai/opal/layouts";
-import { SvgExternalLink, SvgFileText, SvgGlobe } from "@onyx-ai/opal/icons";
-import {
-  SvgBraintrust,
-  SvgConfluence,
-  SvgGithub,
-  SvgGmail,
-  SvgGoogleDrive,
-  SvgHubspot,
-  SvgJira,
-  SvgLinear,
-  SvgNotion,
-  SvgSalesforce,
-  SvgSharepoint,
-  SvgSlack,
-  SvgTeams,
-  SvgZendesk,
-} from "@onyx-ai/opal/logos";
-import type { IconFunctionComponent } from "@onyx-ai/opal/types";
+import { SvgExternalLink } from "@onyx-ai/opal/icons";
 
 import type { CoeditorHandle } from "@/lib/editor/components";
 import { relativeTime } from "@/lib/time";
 import { useIsMobile } from "@/lib/viewport";
-import type { SourceRef, SourceSpan, WriteProvenance } from "@/types";
+import type { SourceRef, SourceSpan } from "@/types";
 
 import { EditorEdgeScrollbar } from "./EditorEdgeScrollbar";
 import { SvgListLines } from "./icons";
 import { PanelSearchField } from "./PanelSearch";
 import { SourceAnchorRail } from "./SourceAnchorRail";
+import { sourceIcon, sourceKey, sourceTypeLabel } from "./sources";
 
 interface Props {
   sources: SourceRef[];
@@ -56,42 +40,6 @@ interface Props {
   /** Fires with the hovered card's source key (null on leave), so the
    * page can light that source's doc spans. */
   onHoverSource?: (key: string | null) => void;
-}
-
-const SOURCE_ICONS: Record<string, IconFunctionComponent> = {
-  braintrust: SvgBraintrust,
-  confluence: SvgConfluence,
-  github: SvgGithub,
-  gmail: SvgGmail,
-  google_drive: SvgGoogleDrive,
-  hubspot: SvgHubspot,
-  jira: SvgJira,
-  linear: SvgLinear,
-  notion: SvgNotion,
-  salesforce: SvgSalesforce,
-  sharepoint: SvgSharepoint,
-  slack: SvgSlack,
-  teams: SvgTeams,
-  web: SvgGlobe,
-  zendesk: SvgZendesk,
-};
-
-export function sourceIcon(type: string | null): IconFunctionComponent {
-  return (type && SOURCE_ICONS[type]) || SvgFileText;
-}
-
-// "google_drive" → "Google Drive", for the connector chip.
-export function sourceTypeLabel(type: string): string {
-  return type
-    .split(/[_-]/)
-    .map((w) => w.charAt(0).toUpperCase() + w.slice(1))
-    .join(" ");
-}
-
-/** Stable card identity, mirroring the backend's dedupe key (document id,
- * falling back to url or title). Also joins a card to its doc spans. */
-export function sourceKey(s: WriteProvenance): string {
-  return s.source_document_id || s.source_url || s.source_title || "";
 }
 
 interface SourceCardProps {
